requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
ollama>=0.4.0
pyyaml>=6.0
